[aml] sync build and rootfs to new m1/m3 aml kernel drivers

+++ package/amlogic/libplayer/src/amplayer/player/include/message.h
@@ -56,8 +56,8 @@ typedef enum {
 } get_info_t;
 
 typedef struct {
-    int min;
-    int max;
+    float min;
+    float max;
 } volume_range_t;
 
 typedef struct {
